requirements.txt
# Core dependencies
mcp>=1.0.0
office365-rest-python-client>=2.5.0
msal>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development dependencies (optional)
# Uncomment to install with: pip install -r requirements.txt -r requirements-dev.txt
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
